Hard Disks Project Documentation

HardDisks.predictCollisionInCell — Method

predictCollisionInCell(box, p, row, col, currTime, currPartner)

Computes the next collision time and partner of particle p against the particles inside the cell at (row,col).

Arguments

box::Box: the box containing the particle configuration and grid

p::Particle: the particle for which the prediction is computed

row::Int: the row position of the cell to check against

col::Int: the column position of the cell to check against

currTime::Float64: the current time of an existing prediction for particle p

currPartner::Int: the current partner of an existing prediction for particle p

Returns

currTime::Float64: the time of the next collision

currPartner::Int: the index of the collision partner

HardDisks.predictCollisionInNeighborhood — Method

predictCollisionInNeighborhood(box, p)

Computes the next collision time and partner of particle p against the particles inside an L-shaped neighborhood:

|   | X | X |
|   | p | X |
|   |   | X |

Arguments

box::Box: the box containing the particle configuration and grid

p::Particle: the particle for which the prediction is computed

Returns

time::Float64: the time of the next collision

partner::Int: the index of the collision partner

HardDisks.predictCollisionInNeighborhood — Method

predictCollisionInNeighborhood(box, p, dy, dx)

Computes the next collision time and partner of particle p against the particles inside a proper neighborhood:

|   |   | X |
|   | p | X |
|   |   | X |

| X | X | X |
|   | p |   |
|   |   |   |

| X | X | X |
|   | p | X |
|   |   | X |

| X | X | X |
| X | X | X |
| X | X | X |

Arguments

box::Box: the box containing the particle configuration and grid

p::Particle: the particle for which the prediction is computed

dy::Int: the direction (sign) of the step the particle took while changing rows

dx::Int: the direction (sign) of the step the particle took while changing columns

Returns

time::Float64: the time of the next collision

partner::Int: the index of the collision partner

HardDisks.handleCollision! — Method

handleCollision!(box, event)

Arguments

box::Box: the box containing the particle configuration, grid and event queue

event::Event: the collision event to handle; it can be a wall or particle collision

HardDisks.updateBox! — Method

updateBox!(box)

Handles the next event in line and updates the event queue and the state of the particles involved.

Arguments

box::Box: the box containing the particle configuration, grid and event queue

HardDisks.updateBox! — Method

updateBox!(box, ϕMax, steps)

Calls updateBox!(box) until the packing fraction reaches an upper limit or its relative increment gets lower than a given threshold

This is the main program loop. The algorithm follows this steps:

Gets the next event from the Indexed Minimum Priority Queue

Handles the event according to its type

For a particle collision both particles get their velocities updated. Only the event of particle p gets updated. The partner gets its event updated to a CHECK

For a wall collision the particle involved gets its velocity and event updated

For a transfer event the particle involved gets its grid position updated

Arguments

box::Box: the box containing the particle configuration, grid and event queue