box.jl Documentation

HardDisks.predictCollisionInCell — Method

predictCollisionInCell(box, p, row, col, currTime, currPartner)

Computes the next collision time and partner of particle p against the particles inside the cell at (row,col).

Arguments

box::Box: the box containing the particle configuration and grid

p::Particle: the particle for which the prediction is computed

row::Int: the row position of the cell to check against

col::Int: the column position of the cell to check against

currTime::Float64: the current time of an existing prediction for particle p

currPartner::Int: the current partner of an existing prediction for particle p

Returns

currTime::Float64: the time of the next collision

currPartner::Int: the index of the collision partner

HardDisks.predictCollisionInNeighborhood — Method

predictCollisionInNeighborhood(box, p)

Computes the next collision time and partner of particle p against the particles inside an L-shaped neighborhood:

|   | X | X |
|   | p | X |
|   |   | X |

Arguments

box::Box: the box containing the particle configuration and grid

p::Particle: the particle for which the prediction is computed

Returns

time::Float64: the time of the next collision

partner::Int: the index of the collision partner